Refactor : layout 구조 리팩토링 - layout이 server component로 작동하도록 수정 - client component인 layoutCore 컴포넌트 구현 - layoutCore 컴포넌트 내에 클라이언트 로직 구현
